Make GIS helpers into methods on Database and Table. Remove gis.py.
